Make chatHooks proposal self contained
